+++ src/event/event.php
@@ -61,6 +61,7 @@ class Event extends Resource
             "issuing-invoice" => Event::issuingInvoiceLogResource(),
             "issuing-purchase" => Event::issuingPurchaseLogResource(),
             "credit-note" => Event::creditNoteLogResource(),
+            "pix-dispute" => Event::pixDisputeLogResource()
         ];
 
         if (!isset($makerOptions[$subscription])) {
@@ -210,6 +211,20 @@ class Event extends Resource
         };
     }
 
+    private static function pixDisputeLogResource()
+    {
+        return function ($array) {
+            $dispute = function ($array) {
+                return new PixDispute($array);
+            };
+            $array["dispute"] = API::fromApiJson($dispute, $array["dispute"]);
+            $log = function ($array) {
+                return new PixDispute\Log($array);
+            };
+            return API::fromApiJson($log, $array);
+        };
+    }
+
     /**
     # Retrieve a specific notification Event
 
@@ -300,8 +315,8 @@ class Event extends Resource
     If isDelivered is true, the event will no longer be returned on queries with isDelivered=false.
 
     ## Parameters (required):
-        - id [array of strings]: Event unique ids. ex: "5656565656565656"
-        - isDelivered [bool]: If true and event hasn't been delivered already, event will be set as delivered. ex: true
+        - id [string]: Event unique id. ex: "5656565656565656"
+        - options [array]: Options to update the Event. ex: ["isDelivered" => true]
 
     ## Parameters (optional):
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra/Settings::setUser() was used before function call
@@ -309,9 +324,9 @@ class Event extends Resource
     ## Return:
         - target Event with updated attributes
      */
-    public static function update($id, $isDelivered, $user = null)
+    public static function update($id, $options, $user = null)
     {
-        return Rest::patchId($user, Event::resource(), $id, [$isDelivered]);
+        return Rest::patchId($user, Event::resource(), $id, $options);
     }
 
     /**