+++ src/utils/rest.php
@@ -8,7 +8,9 @@ use StarkCore\Utils\StarkHost;
 class Rest
 {
     private const apiVersion = "v2";
-    private const sdkVersion = "0.13.2";
+
+    private const sdkVersion = "0.16.0";
+
     private const host = StarkHost::infra;
 
     public static function getPage($user, $resource, array $query = [])